% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freepst-package.R
\docType{package}
\name{freepst}
\alias{freepst}
\title{Read OST/PST Files}
\description{
An 'rJava' wrapper around 'java-libpst' \url{https://github.com/rjohnsondev/java-libpst}.
}
\author{
Bob Rudis (bob@rud.is)
}
