% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-folder.R
\name{read_folder}
\alias{read_folder}
\title{Get all messages in an OST/PST with nested folder metadata}
\usage{
read_folder(path)
}
\arguments{
\item{path}{path to OST/PST}
}
\value{
data frame
}
\description{
Get all messages in an OST/PST with nested folder metadata
}
\examples{
read_folder(system.file("extdata/example-2013.ost", package="freepst"))
read_folder(system.file("extdata/dist-list.pst", package="freepst"))
}
\author{
Torben Tvedebrink
}
