% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freepst-package.R
\docType{package}
\name{freepst}
\alias{freepst}
\title{Read OST/PST Files}
\description{
An 'rJava' wrapper around 'java-libpst' \url{https://github.com/rjohnsondev/java-libpst}.
The PST File format is used by Outlook for the storage of emails. Over the
years many people have accumulated a large amount of important email and
other information into these files. Tools are provided to enable folks to
access and extract this information for archival or analysis.
}
\author{
Bob Rudis (bob@rud.is)
}
